+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 2.x branch
 ## 2.15 branch
 ### 2.15.0
-* `BackupTrait` had been deprecated and has been removed.
+* `BackupTrait` had been deprecated and has been removed;
+* `AbstractBackupUtility::getDriver()` had been deprecated and has been removed.
 
 ## 2.14 branch
 ### 2.14.3